feat: implement api_root view to provide API endpoints for users, teams, activity, leaderboard, and workouts
